% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bikecourse_function.R
\name{bikecourse_function}
\alias{bikecourse_function}
\title{Bikecourse function}
\usage{
bikecourse_function(x, y, data, threshold, xlab, ylab, yax, size)
}
\arguments{
\item{x}{numeric variable x (written as data$x)}

\item{y}{variable y (written as data$y)}

\item{data}{name of the dataframe}

\item{threshold}{threshold for significance (default is 0.05)}

\item{xlab}{name of the x-axis of boxplot (default is x)}

\item{ylab}{name of the y-axis of boxplot (default is y)}

\item{yax}{location of significance letters on the y-axis}

\item{size}{size of the significance letters (default is 7)}
}
\description{
Findin association between a numeric variable x and a variable y with a boxplot (with significance letters) or a scatterplot (with regressionline) as output
}
