% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dunn_function.R
\name{dunn_function}
\alias{dunn_function}
\title{Dunn graph}
\usage{
dunn_function(x, y, threshold = 0.05, xlab = "x", ylab = "y", yax, size = 7)
}
\arguments{
\item{x}{numeric variable x (written as data$x)}

\item{y}{variable y (written as data$y)}

\item{threshold}{threshold for significance (default is 0.05)}

\item{xlab}{name of the x-axis of boxplot (default is x)}

\item{ylab}{name of the y-axis of boxplot (default is y)}

\item{yax}{location of significance letters on the y-axis}

\item{size}{size of the significance letters (default is 7)}
}
\description{
Boxplot between a numeric variable x and a factor variable y with difference between groups based on dunn test
}
