% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_function.R
\name{ggplot_function}
\alias{ggplot_function}
\title{Scatterplot with regressionline}
\usage{
ggplot_function(x, y, xlab, ylab, data)
}
\arguments{
\item{x}{numeric variable x (should be written as data$x)}

\item{y}{variable y (should be written as data$y)}

\item{data}{name of the dataframe}
}
\description{
Scatterplot of a numeric variable x and a numeric variable y with a linear regressionline
}
