% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Kruskal_function.R
\name{Kruskal_function}
\alias{Kruskal_function}
\title{Kruskal wallis}
\usage{
Kruskal_function(x, y, data)
}
\arguments{
\item{x}{numeric variable x (should be written as data$x)}

\item{y}{variable y (should be written as data$y)}

\item{data}{name of the dataframe}
}
\description{
Kruskal wallis test between a numeric variable x and a factor variable y
}
