% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spearman_function.R
\name{Spearman_function}
\alias{Spearman_function}
\title{Spearman rank}
\usage{
Spearman_function(x, y)
}
\arguments{
\item{x}{numeric variable x (should be written as data$x)}

\item{y}{variable y (should be written as data$y)}
}
\description{
Spearman rank test between a numeric variable x and a numeric variable y
}
